agent/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
anthropic>=0.42.0
pydantic>=2.0
pydantic-settings>=2.0
click>=8.0
Pillow>=10.0
python-dotenv>=1.0
httpx>=0.27.0
